+++ src/models/query_parameters.rs
@@ -2,9 +2,9 @@ use serde::Deserialize;
 
 #[derive(Deserialize)]
 pub struct QueryParameters {
-    pub from: Option<i64>,
-    pub count: Option<i32>,
+    pub from: Option<f64>,
+    pub count: Option<f64>,
     pub interval: Option<String>,
-    pub to: Option<i64>,
+    pub to: Option<f64>,
     pub pool: Option<String>,
 }
